server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
beanie==1.25.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cloudinary==1.36.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
filetype==1.2.0
email-validator==2.1.0.post1
pymongo<4.6

